refactored parser_utils
- Split into two higher level small modules: parsers and predicates
