+++ src/parse/mod.rs
@@ -1,7 +1,8 @@
 pub mod ast;
 pub mod input;
 mod lines;
-mod parser_utils;
+pub mod parsers;
+pub mod predicates;
 pub mod segment;
 #[cfg(test)]
 pub mod test_utils;
