--- src/parse/parser_utils/mod.rs
@@ -1,6 +1,0 @@
-//! Module containing utilities extending the [crate::parse::parser] framework.
-mod parsers;
-mod predicates;
-
-pub use parsers::*;
-pub use predicates::*;
