+++ src/parse/segment/fenced_code/backticks.rs
@@ -1,5 +1,5 @@
 use crate::Segment;
-use crate::parse::parser_utils::{indented_by_less_than_4, line_ending_or_eof, space_or_tab};
+use crate::parse::parsers::{indented_by_less_than_4, line_ending_or_eof, space_or_tab};
 use crate::parse::traits::ParseLine;
 use parser::{Map, ParseResult, Parser, consumed};
